requirements.txt
flask
flask_socketio
numpy
Pillow
openai
paddlex
keyboard
pyautogui
pyinstaller